requirements.txt
flask
flask_sqlalchemy
requests
flasgger
PyJWT
flask_migrate
event_bus